requirements.txt
Flask
pandas
yagmail
google-auth
gspread
schedule
